Fix for exception being through in PinPoint Node
pinpoint_node was throwing an exception when attempting to convert a time stamp in onGlobalPoseChangedHandler

This was being caused by a missed break statement in the pinpoint_localization_client code. A status event was being fired
but fall through would cause an onGlobalPoseChanged to be emited as well.

Added addition exception messages and cleaned up the code surround the bug.