Remove URDF usage from driver (#33)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description

<!--- Describe your changes in detail -->
The pin_point_application node attempts to conduct a transform between
the 'pinpoint' and the 'base_link' frames, and this transform was
initially described in the robot.urdf file stored in this repository
(though the transform was never published by robot_state_publisher). To
remain consistent with how urdf files are used within CARMA Platform,
this urdf file has been removed and will be stored in a vehicle
calibration folder for each CARMA Messenger-equipped vehicle (currently
only the 2018 Chevrolet Tahoe). CARMA Messenger will load this urdf file
at launch and the transform will be published by robot_state_publisher
to the /tf_static topic.

Related carma-messenger PR: [PR
#186](https://github.com/usdot-fhwa-stol/carma-messenger/pull/186)
Related carma-vehicle-calibration PR: [PR
#113](https://github.com/usdot-fhwa-stol/carma-vehicle-calibration/pull/113)

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->
[CARMA Messenger Issue
#187](https://github.com/usdot-fhwa-stol/carma-messenger/issues/187)

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
This update is required in support of publishing the urdf information to
the /tf_static topic (see related PRs) so that this driver can properly
process velocity data and publish it to the /position/velocity topic.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Local integration test

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.